ShangOnline/requirements.txt
diff-match-patch==20121119
Django==1.11.6
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-ranged-response==0.2.0
django-reversion==2.0.13
django-simple-captcha==0.5.7
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
jdcal==1.4
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.1.0
pytz==2018.4
PyYAML==3.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
